pyproject.toml
[tool.poetry]
name = "final_proj"
version = "0.1.0"
description = ""
authors = ["yordankitov <yordankitov@infinitelambda.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
snowflake-connector-python = "^2.7.4"
boto3 = "^1.21.11"
pandas = "^1.4.1"
snowflake-sqlalchemy = "^1.3.3"
pyarrow = "6.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

